work

+++ common/gui/widgets/textblock.cpp
@@ -3,7 +3,7 @@
 #include "button.h"
 #include "checkbox.h"
 #include "editbox.h"
-#include "dropdowns.h"
+#include "droplist.h"
 #include "image.h"
 #include "insdraw.h"
 #include "link.h"
@@ -35,5 +35,5 @@ void TextBlock::changevalue(const char* newv)
 
 int TextBlock::square()
 {
-	return g_font[m_font].gheight;
+	return (int)g_font[m_font].gheight;
 }
